fix: XYNE-17394 address flow board follow-ups
