+++ apps/backend/src/services/tickets/kanbanQueryBuilder.ts
@@ -1,5 +1,11 @@
 import { Prisma } from '@prisma/client';
-import { TicketPriority, TicketStatusV2, UserResponsibility } from '@xyne/shared';
+import {
+  FLOW_STEP_ROOT_ID_FIELD,
+  TicketPriority,
+  TicketStatusV2,
+  UserResponsibility,
+  type FlowStepVisibilityOptions,
+} from '@xyne/shared';
 import { parseAssigneeFilter } from '@xyne/shared/zero/queries';
 
 const SUPPORT_TICKET_TYPE = 'Support';
@@ -42,7 +48,7 @@ export type KanbanTicketFilters = {
   dynamicFields?: Record<string, string[] | { start?: number; end?: number }>;
 };
 
-export type KanbanTicketQueryContext = {
+export type KanbanTicketQueryContext = FlowStepVisibilityOptions & {
   workspaceId: string;
   currentUserId?: string;
   viewMode: KanbanTicketViewMode;
@@ -211,6 +217,7 @@ export const buildKanbanTicketWhere = (
       { isArchived: false },
       buildChannelAccessFilter(context.currentUserId),
       buildScopeFilter(context),
+      context.excludeFlowSteps ? { [FLOW_STEP_ROOT_ID_FIELD]: null } : undefined,
       {
         OR: [{ ticketType: null }, { ticketType: { not: SUPPORT_TICKET_TYPE } }],
       },
