+++ apps/dashboard/src/components/Board/FlowPlanEditor/FlowPlanEditor.tsx
@@ -2,15 +2,11 @@ import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import ReactFlow, {
   Background,
   BackgroundVariant,
-  Handle,
-  Position,
   useNodesState,
   type Connection,
   type Edge,
   type Node,
-  type NodeProps,
   type ReactFlowInstance,
-  MarkerType,
   MiniMap,
   Panel,
   SelectionMode,
@@ -18,28 +14,11 @@ import ReactFlow, {
 import 'reactflow/dist/style.css';
 import { v4 as uuidv4 } from 'uuid';
 import { toast } from 'sonner';
-import {
-  AlertTriangle,
-  Boxes,
-  ChevronRight,
-  Plus,
-  Trash2,
-  LayoutGrid,
-  Ticket,
-  Settings2,
-  FileText,
-  UserCheck,
-  UserPlus,
-  Search,
-  SquareDashedMousePointer,
-  X,
-  GitFork,
-} from 'lucide-react';
+import { Plus, LayoutGrid, FileText, UserCheck, SquareDashedMousePointer, X } from 'lucide-react';
 import {
   FormContextType,
   FormEntityType,
   FormFieldType,
-  FlowPlanModel,
   type FlowPlanGroup,
   type FlowPlanDecision,
   type FlowPlanNode,
@@ -49,48 +28,18 @@ import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
-  DropdownMenuSub,
-  DropdownMenuSubContent,
-  DropdownMenuSubTrigger,
   DropdownMenuTrigger,
 } from '../../ui/dropdown-menu';
-import Textarea from '../../ui/Textarea';
-import { Button } from '../../ui/Button';
 import { STATUS_OPTIONS } from '../BoardStageConfigScreen/BoardStageConfigScreen.types';
-import { Popover } from '../../ui/Popover/Popover';
-import UserAvatar, { AvatarSize } from '../../UserAvatar/UserAvatar';
-import { useActiveUsers } from '../../../hooks/useUsers';
-import { getUserDisplayName } from '../../../utils/userDisplayName';
 import { cn } from '../../../utils/classNames';
-import { TransitionFormPicker } from '../TransitionFormPicker/TransitionFormPicker';
-import { CreateFormSlideOut } from '../CreateFormSlideOut/CreateFormSlideOut';
-import type { FormField } from '../CreateFormSlideOut/CreateFormSlideOut.types';
 import { formService } from '../../../services/Form/formService';
 import {
-  getFieldTypeLabel,
   mapFormDetailsToBuilderFields,
   mapFormFieldsToApiPayload,
 } from '../../../utils/board/formFieldApiMapper';
 import { useCachedQuery } from '../../../hooks/useCachedQuery';
 import { queries } from '../../../zero/queries';
-import {
-  computeFlowLayout,
-  computeGroupInternalLayout,
-  getFlowPlanEditorIssues,
-  VIRTUAL_ROOT_ID,
-  CARD_WIDTH,
-  type FlowLayoutItem,
-} from './FlowPlanEditor.utils';
-import {
-  FlowGroupNode,
-  GROUP_ENTRY_HANDLE,
-  GROUP_EXIT_HANDLE,
-  GROUP_INPUT_HANDLE,
-  GROUP_OUTPUT_HANDLE,
-  collapsedGroupCoverHeight,
-  flowGroupColor,
-  type FlowGroupNodeData,
-} from '../FlowRun/FlowGroupNode';
+import { getFlowPlanEditorIssues, VIRTUAL_ROOT_ID, CARD_WIDTH } from './FlowPlanEditor.utils';
 import { FlowPlanEdge } from './FlowPlanEdge';
 import { FlowDecisionConfigPanel } from './FlowDecisionConfigPanel';
 import { disconnectDecisionTargets, reassignDecisionRoute } from './flowDecisionMutations';
@@ -116,21 +65,23 @@ import {
   withoutStaleDecisionParents,
   type PlanState,
 } from './flowPlanGraph';
+import {
+  DEFAULT_GATE,
+  FLOW_PLAN_NODE_TYPES,
+  gateOf,
+  type AddStepKind,
+  type EditorNodeData,
+} from './FlowPlanNodes';
+import { useFlowPlanEdges } from './useFlowPlanEdges';
+import { useBuildFlowPlanNodes } from './useBuildFlowPlanNodes';
+import {
+  FlowStepConfigPanel,
+  type ConfirmationGate,
+  type FormBuilderData,
+  type FormGate,
+} from './FlowStepConfigPanel';
 
-interface FormBuilderData {
-  formName: string;
-  formDescription: string;
-  fields: FormField[];
-}
-
-type ConfirmationGate = Extract<FlowStepGate, { type: 'confirmation' }>;
-type FormGate = Extract<FlowStepGate, { type: 'form' }>;
-
-export const DEFAULT_GATE: FlowStepGate = { type: 'confirmation' };
-
-export function gateOf(node: FlowPlanNode): FlowStepGate {
-  return node.gate ?? DEFAULT_GATE;
-}
+export { gateOf } from './FlowPlanNodes';
 
 const STEP_CARD_HEIGHT = 170;
 
@@ -155,490 +106,7 @@ const groupSelectionError = (reason: CreateGroupReason): string => {
   }
 };
 
-type AddStepKind = FlowStepGate['type'] | 'group' | 'decision';
-
-interface PlanNodeData {
-  planNode: FlowPlanNode | null; // null = the virtual "Main ticket" root
-  detached: boolean;
-  onUpdate: (patch: Partial<FlowPlanNode>) => void;
-  onDelete: () => void;
-  onAddStep: (kind: AddStepKind) => void;
-  onConfigure: () => void;
-  configuring: boolean;
-  readOnly: boolean;
-  canAddGroup: boolean;
-  canAddDecision: boolean;
-  routesThroughDecision: boolean;
-  validationWarning?: string;
-}
-
-const AssigneeSummary: React.FC<{ userId: string | null }> = ({ userId }) => {
-  const users = useActiveUsers();
-  const assigneeId = userId?.replace(/^(user:|group:)/, '') || '';
-  if (!assigneeId) return <span className='text-muted-foreground'>Unassigned</span>;
-  const user = users?.find(candidate => candidate.id === assigneeId);
-  return (
-    <span className='flex items-center gap-1.5 min-w-0 font-medium text-foreground'>
-      <UserAvatar userId={assigneeId} showActiveStatus={false} size={AvatarSize.SM} />
-      <span className='truncate'>{user ? getUserDisplayName(user) : '…'}</span>
-    </span>
-  );
-};
-
-const StepAssigneePicker: React.FC<{
-  value: string | null;
-  onChange: (userId: string | null) => void;
-}> = ({ value, onChange }) => {
-  const [open, setOpen] = useState(false);
-  const [search, setSearch] = useState('');
-  const users = useActiveUsers();
-
-  const assigneeId = value?.replace(/^(user:|group:)/, '') || '';
-  const filteredUsers = useMemo(() => {
-    if (!users) return [];
-    const query = search.trim().toLowerCase();
-    if (!query) return users;
-    return users.filter(user => {
-      const name = getUserDisplayName(user).toLowerCase();
-      const email = (user.email ?? '').toLowerCase();
-      return name.includes(query) || email.includes(query);
-    });
-  }, [users, search]);
-
-  const pick = (userId: string | null): void => {
-    onChange(userId);
-    setOpen(false);
-    setSearch('');
-  };
-
-  const avatar = assigneeId ? (
-    <UserAvatar userId={assigneeId} showActiveStatus={false} size={AvatarSize.SM} />
-  ) : (
-    <span className='inline-flex items-center justify-center w-5 h-5 rounded-sm border border-dashed border-border text-muted-foreground hover:text-foreground hover:border-foreground'>
-      <UserPlus className='w-3 h-3' />
-    </span>
-  );
-
-  const assignee = users?.find(candidate => candidate.id === assigneeId);
-
-  return (
-    <Popover
-      trigger={
-        <button
-          type='button'
-          onClick={e => {
-            e.stopPropagation();
-            setOpen(prev => !prev);
-          }}
-          className='flex w-full items-center gap-2 rounded-lg border border-border bg-background px-3 py-2 text-[12px] hover:bg-muted transition-colors'
-          title={assigneeId ? 'Change default assignee' : 'Set default assignee for this step'}
-          data-track-category='flow_plan_editor'
-          data-track-name='toggle_step_assignee'
-        >
-          {avatar}
-          <span
-            className={`flex-1 min-w-0 truncate text-left ${
-              assigneeId ? 'text-foreground' : 'text-muted-foreground'
-            }`}
-          >
-            {assigneeId ? (assignee ? getUserDisplayName(assignee) : '…') : 'Unassigned'}
-          </span>
-        </button>
-      }
-      open={open}
-      onOpenChange={setOpen}
-      modal
-      align='end'
-      sideOffset={4}
-      className='p-0 w-64 z-[9999]'
-    >
-      <div className='flex flex-col max-h-72'>
-        <div className='p-2 border-b border-border'>
-          <div className='relative'>
-            <Search className='absolute left-2.5 top-1/2 -translate-y-1/2 w-3.5 h-3.5 text-muted-foreground' />
-            <input
-              type='text'
-              placeholder='Search users...'
-              value={search}
-              onChange={e => setSearch(e.target.value)}
-              className='w-full pl-8 pr-2 py-1.5 border border-input rounded-md bg-background text-xs text-foreground placeholder:text-muted-foreground focus:ring-2 focus:ring-blue-500 focus:border-blue-500 outline-none'
-              data-track-category='flow_plan_editor'
-              data-track-name='search_step_assignee'
-            />
-          </div>
-        </div>
-        <div className='overflow-y-auto flex-1'>
-          <button
-            type='button'
-            onClick={e => {
-              e.stopPropagation();
-              pick(null);
-            }}
-            className={cn(
-              'w-full text-left px-3 py-2 text-xs hover:bg-muted transition-colors flex items-center gap-2',
-              !assigneeId && 'bg-muted',
-            )}
-            data-track-category='flow_plan_editor'
-            data-track-name='clear_step_assignee'
-          >
-            <span className='flex items-center justify-center w-5 h-5 rounded-sm bg-border'>
-              <X className='w-3 h-3 text-muted-foreground' />
-            </span>
-            <span className='text-foreground'>Unassigned</span>
-          </button>
-          {filteredUsers.map(user => (
-            <button
-              key={user.id}
-              type='button'
-              onClick={e => {
-                e.stopPropagation();
-                pick(user.id);
-              }}
-              className={cn(
-                'w-full text-left px-3 py-2 text-xs hover:bg-muted transition-colors flex items-center gap-2',
-                assigneeId === user.id && 'bg-muted',
-              )}
-              data-track-category='flow_plan_editor'
-              data-track-name='select_step_assignee'
-            >
-              <UserAvatar userId={user.id} showActiveStatus={false} size={AvatarSize.SM} />
-              <div className='flex-1 min-w-0'>
-                <div className='text-foreground truncate'>{getUserDisplayName(user)}</div>
-                {user.email ? (
-                  <div className='text-[10px] text-muted-foreground truncate'>{user.email}</div>
-                ) : null}
-              </div>
-            </button>
-          ))}
-          {filteredUsers.length === 0 && (
-            <div className='px-3 py-3 text-xs text-muted-foreground text-center'>
-              No users found
-            </div>
-          )}
-        </div>
-      </div>
-    </Popover>
-  );
-};
-
-const PlanNodeCard: React.FC<NodeProps<PlanNodeData>> = ({ data, selected }) => {
-  const { planNode, onUpdate, onDelete, onAddStep, onConfigure, readOnly } = data;
-  const isRoot = planNode === null;
-  const warning =
-    data.validationWarning ??
-    (data.detached ? 'Not connected to the flow — connect it before saving' : undefined);
-
-  return (
-    <div
-      className={`w-[240px] bg-background rounded-[10px] border-2 shadow-[0px_2px_8px_0px_rgba(5,5,6,0.07)] transition-all ${
-        selected
-          ? 'border-[#6276be]'
-          : warning
-            ? 'border-amber-400'
-            : isRoot
-              ? 'border-[#6276be]/50'
-              : 'border-border'
-      }`}
-    >
-      {!isRoot && (
-        <Handle
-          type='target'
-          position={Position.Top}
-          isConnectable={!readOnly}
-          className='!w-3 !h-3 !bg-[#6276be] !border-2 !border-background !rounded-full'
-          style={{ top: -7 }}
-        />
-      )}
-      <Handle
-        type='source'
-        position={Position.Bottom}
-        isConnectable={!readOnly}
-        className='!w-3 !h-3 !bg-[#6276be] !border-2 !border-background !rounded-full'
-        style={{ bottom: -7 }}
-      />
-
-      <div className='flex items-center justify-between px-3 py-2 border-b border-border'>
-        <span className='flex items-center gap-1.5 text-[11px] font-semibold text-muted-foreground uppercase tracking-[0.5px]'>
-          <Ticket size={12} className='text-[#6276be]' />
-          {isRoot ? 'Main ticket' : 'Sub ticket'}
-        </span>
-        {!isRoot && (
-          <div className='flex items-center gap-1 nodrag' onPointerDown={e => e.stopPropagation()}>
-            {warning && (
-              <span
-                className='flex items-center text-amber-500'
-                title={warning}
-                aria-label={warning}
-              >
-                <AlertTriangle size={12} />
-              </span>
-            )}
-            {!readOnly && (
-              <>
-                <button
-                  type='button'
-                  onClick={onConfigure}
-                  data-track-category='flow_plan_editor'
-                  data-track-name='configure_step_gate'
-                  className={`p-1 rounded-md transition-colors ${
-                    data.configuring
-                      ? 'bg-[#6276be] text-white'
-                      : 'hover:bg-muted text-muted-foreground hover:text-[#6276be]'
-                  }`}
-                  title='Configure what this step waits on'
-                >
-                  <Settings2 size={12} />
-                </button>
-                <button
-                  type='button'
-                  onClick={onDelete}
-                  data-track-category='flow_plan_editor'
-                  data-track-name='delete_step'
-                  className='p-1 rounded-md hover:bg-red-50 text-muted-foreground hover:text-red-500 transition-colors'
-                  title='Delete step'
-                >
-                  <Trash2 size={12} />
-                </button>
-              </>
-            )}
-          </div>
-        )}
-      </div>
-
-      <div className='px-3 py-3'>
-        {isRoot ? (
-          <p className='text-[12px] text-muted-foreground leading-[18px]'>
-            Each run starts from a main ticket. Steps below it are created as the run progresses.
-          </p>
-        ) : (
-          <>
-            <input
-              type='text'
-              value={planNode.title}
-              readOnly={readOnly}
-              onChange={e => onUpdate({ title: e.target.value })}
-              placeholder='Step title...'
-              data-track-category='flow_plan_editor'
-              data-track-name='input_step_title'
-              className='nodrag w-full text-[13px] font-medium text-foreground bg-transparent border-none focus:outline-none'
-              onPointerDown={e => e.stopPropagation()}
-            />
-            <div className='mt-2 flex flex-col gap-1 text-[11px]'>
-              <div className='flex items-center justify-between gap-2'>
-                <span className='text-muted-foreground'>Waits for</span>
-                <span className='flex items-center gap-1 font-medium text-foreground'>
-                  {gateOf(planNode).type === 'form' ? (
-                    <>
-                      <FileText size={11} className='text-[#6276be]' />
-                      Form
-                    </>
-                  ) : (
-                    <>
-                      <UserCheck size={11} className='text-[#6276be]' />
-                      Confirmation
-                    </>
-                  )}
-                </span>
-              </div>
-              <div className='flex items-center justify-between gap-2'>
-                <span className='text-muted-foreground'>Assignee</span>
-                <AssigneeSummary userId={planNode.assignedTo ?? null} />
-              </div>
-            </div>
-          </>
-        )}
-      </div>
-
-      {!readOnly && !data.routesThroughDecision && (
-        <div className='px-3 pb-2.5'>
-          <DropdownMenu>
-            <DropdownMenuTrigger
-              className='nodrag flex items-center gap-1.5 w-full justify-center border border-dashed border-[#6276be]/50 hover:border-[#6276be] rounded-lg px-2 py-1 text-[11px] text-[#6276be] hover:text-[#4f61a8] font-medium transition-colors outline-none'
-              onPointerDown={e => e.stopPropagation()}
-              data-track-category='flow_plan_editor'
-              data-track-name='add_step'
-            >
-              <Plus size={11} />
-              Add step
-            </DropdownMenuTrigger>
-            {/* Offset clears the card's bottom padding + border so the menu
-                doesn't overlap the card */}
-            <DropdownMenuContent align='center' sideOffset={18} className='z-[9999]'>
-              {data.canAddGroup && (
-                <DropdownMenuItem
-                  onSelect={() => onAddStep('group')}
-                  className='flex items-center gap-2'
-                >
-                  <Boxes size={13} className='text-[#8b5cf6]' />
-                  <div className='flex flex-col'>
-                    <span className='text-[12px] font-medium'>Group section</span>
-                    <span className='text-[11px] text-muted-foreground'>
-                      A sub-flow that acts as one step of the plan
-                    </span>
-                  </div>
-                </DropdownMenuItem>
-              )}
-              <DropdownMenuSub>
-                <DropdownMenuSubTrigger className='flex items-center gap-2'>
-                  <Ticket size={13} className='text-[#6276be]' />
-                  <div className='flex flex-col flex-1'>
-                    <span className='text-[12px] font-medium'>Step</span>
-                    <span className='text-[11px] text-muted-foreground'>
-                      One ticket that waits at a gate
-                    </span>
-                  </div>
-                  <ChevronRight size={13} className='shrink-0 text-muted-foreground' />
-                </DropdownMenuSubTrigger>
-                <DropdownMenuSubContent sideOffset={10} className='z-[9999]'>
-                  <DropdownMenuItem
-                    onSelect={() => onAddStep('confirmation')}
-                    className='flex items-center gap-2'
-                  >
-                    <UserCheck size={13} className='text-[#6276be]' />
-                    <div className='flex flex-col'>
-                      <span className='text-[12px] font-medium'>Waits for confirmation</span>
-                      <span className='text-[11px] text-muted-foreground'>
-                        Someone confirms before it completes
-                      </span>
-                    </div>
-                  </DropdownMenuItem>
-                  <DropdownMenuItem
-                    onSelect={() => onAddStep('form')}
-                    className='flex items-center gap-2'
-                  >
-                    <FileText size={13} className='text-[#6276be]' />
-                    <div className='flex flex-col'>
-                      <span className='text-[12px] font-medium'>Waits for a form</span>
-                      <span className='text-[11px] text-muted-foreground'>
-                        Someone fills a form before it completes
-                      </span>
-                    </div>
-                  </DropdownMenuItem>
-                </DropdownMenuSubContent>
-              </DropdownMenuSub>
-              {data.canAddDecision && (
-                <DropdownMenuItem
-                  onSelect={() => onAddStep('decision')}
-                  className='flex items-center gap-2'
-                >
-                  <GitFork size={13} className='text-amber-600' />
-                  <div className='flex flex-col'>
-                    <span className='text-[12px] font-medium'>Decision</span>
-                    <span className='text-[11px] text-muted-foreground'>
-                      Route using a required form field
-                    </span>
-                  </div>
-                </DropdownMenuItem>
-              )}
-            </DropdownMenuContent>
-          </DropdownMenu>
-        </div>
-      )}
-    </div>
-  );
-};
-
-interface DecisionNodeData {
-  decision: FlowPlanDecision;
-  onConfigure: () => void;
-  onDelete: () => void;
-  configuring: boolean;
-  readOnly: boolean;
-  validationWarning?: string;
-}
-
-const DecisionNodeCard: React.FC<NodeProps<DecisionNodeData>> = ({ data, selected }) => (
-  <div
-    className={cn(
-      'relative w-[240px] rounded-xl border-2 bg-[hsl(var(--flow-decision-bg))] shadow-[0_2px_8px_rgba(5,5,6,0.07)]',
-      selected || data.configuring
-        ? 'border-amber-500'
-        : data.validationWarning
-          ? 'border-red-400'
-          : 'border-[hsl(var(--flow-decision-border))]',
-    )}
-  >
-    <Handle
-      id='decision'
-      type='target'
-      position={Position.Top}
-      isConnectable={false}
-      className='!h-3 !w-3 !rounded-full !border-2 !border-background !bg-amber-500'
-      style={{ top: -7 }}
-    />
-    <div className='flex items-center justify-between border-b border-[hsl(var(--flow-decision-divider))] px-3 py-2'>
-      <span className='flex items-center gap-1.5 text-[11px] font-semibold uppercase tracking-[0.5px] text-[hsl(var(--flow-decision-fg))]'>
-        <GitFork size={12} /> Decision
-      </span>
-      <div className='flex items-center gap-1 nodrag'>
-        {data.validationWarning && (
-          <span
-            className='flex items-center text-amber-600'
-            title={data.validationWarning}
-            aria-label={data.validationWarning}
-          >
-            <AlertTriangle size={12} />
-          </span>
-        )}
-        {!data.readOnly && (
-          <>
-            <button
-              type='button'
-              onClick={data.onConfigure}
-              data-track-category='flow_plan_editor'
-              data-track-name='configure_decision'
-              className='rounded p-1 hover:bg-[hsl(var(--flow-decision-hover))]'
-            >
-              <Settings2 size={12} />
-            </button>
-            <button
-              type='button'
-              onClick={data.onDelete}
-              data-track-category='flow_plan_editor'
-              data-track-name='delete_decision'
-              className='rounded p-1 hover:bg-red-50 hover:text-red-500'
-            >
-              <Trash2 size={12} />
-            </button>
-          </>
-        )}
-      </div>
-    </div>
-    <div className='px-3 py-3'>
-      <p className='truncate text-[12px] font-semibold text-foreground'>
-        {data.decision.fieldName || 'Choose a required field'}
-      </p>
-      {data.decision.fieldType === 'STRING' && (
-        <p className='mt-1 text-[10px] text-muted-foreground'>
-          {`${data.decision.operator === 'notEquals' ? 'Does not equal' : 'Equals'} “${data.decision.comparisonValue ?? ''}”`}
-        </p>
-      )}
-      {data.decision.routes.map((route, index) => (
-        <Handle
-          key={route.key}
-          id={route.key}
-          type='source'
-          position={Position.Bottom}
-          isConnectable={!data.readOnly}
-          className='!h-2.5 !w-2.5 !border-2 !border-background !bg-amber-500'
-          style={{
-            left: `${((index + 1) / (data.decision.routes.length + 1)) * 100}%`,
-            bottom: -6,
-          }}
-        />
-      ))}
-    </div>
-  </div>
-);
-
-const NODE_TYPES = {
-  flowPlanNode: PlanNodeCard,
-  flowGroupNode: FlowGroupNode,
-  flowDecisionNode: DecisionNodeCard,
-};
 const EDGE_TYPES = { flowPlan: FlowPlanEdge };
-
-type EditorNodeData = PlanNodeData | FlowGroupNodeData | DecisionNodeData;
 
 export interface FlowPlanEditorProps {
   planNodes: FlowPlanNode[];
@@ -1615,346 +1083,28 @@ export const FlowPlanEditor: React.FC<FlowPlanEditorProps> = ({
    * and user-resized cover sizes are preserved; new/changed nodes fall back
    * to the computed layout.
    */
-  const buildGraphNodes = useCallback(
-    (prev: Node<EditorNodeData>[], relayout: boolean): Node<EditorNodeData>[] => {
-      const plan = planNodesRef.current;
-      const collapsed = collapsedGroupsRef.current;
-      const prevById = new Map(prev.map(n => [n.id, n]));
-
-      // Read-model over the live editor arrays (already normalized in state).
-      const model = new FlowPlanModel({
-        version: 2,
-        updatedAt: 0,
-        nodes: plan,
-        groups: groupsRef.current,
-        decisions: decisionsRef.current,
-      });
-      const activeGroups = model.activeGroups;
-      const groupIds = model.activeGroupIds;
-      const nestedGroups = activeGroups.filter(group => !!group.groupId);
-      const topGroups = activeGroups.filter(group => !group.groupId);
-      const internalByGroup = new Map(
-        nestedGroups.map(group => [
-          group.id,
-          computeGroupInternalLayout(
-            model.membersOf(group.id).map(member => ({
-              ...member,
-              order: model.layoutOrder(member.id),
-            })),
-            STEP_CARD_HEIGHT,
-            model.decisions.filter(
-              decision => model.getNode(decision.parentNodeId)?.groupId === group.id,
-            ),
-          ),
-        ]),
-      );
-      // Cover size always follows the content (collapsed = member list card)
-      const coverSizeOf = (groupId: string): { width: number; height: number } => {
-        if (collapsed.has(groupId)) {
-          return {
-            width: CARD_WIDTH,
-            height: collapsedGroupCoverHeight(
-              model.membersOf(groupId).length + model.childGroupsOf(groupId).length,
-            ),
-          };
-        }
-        const content = internalByGroup.get(groupId);
-        return content
-          ? { width: content.width, height: content.height }
-          : { width: CARD_WIDTH, height: STEP_CARD_HEIGHT };
-      };
-
-      for (const group of topGroups) {
-        const children = model.childGroupsOf(group.id).filter(child => groupIds.has(child.id));
-        const childIds = new Set(children.map(child => child.id));
-        internalByGroup.set(
-          group.id,
-          computeGroupInternalLayout(
-            model.membersOf(group.id).map(member => ({
-              ...member,
-              order: model.layoutOrder(member.id),
-              parentIds: member.parentIds.map(parentId =>
-                childIds.has(parentId) ? coverId(parentId) : parentId,
-              ),
-            })),
-            STEP_CARD_HEIGHT,
-            model.decisions.filter(
-              decision => model.getNode(decision.parentNodeId)?.groupId === group.id,
-            ),
-            100,
-            children.map(child => ({
-              id: coverId(child.id),
-              parentIds: child.parentIds.map(parentId =>
-                childIds.has(parentId) ? coverId(parentId) : parentId,
-              ),
-              order: model.layoutOrder(child.id),
-              ...coverSizeOf(child.id),
-            })),
-          ),
-        );
-      }
-
-      // Top-level layout: ungrouped steps + covers as one block each
-      const layoutItems: FlowLayoutItem[] = [
-        ...plan
-          .filter(step => !step.groupId)
-          .map(step => ({
-            id: step.id,
-            parentIds: step.parentIds.map(parentId =>
-              groupIds.has(parentId) ? coverId(parentId) : parentId,
-            ),
-            order: model.layoutOrder(step.id),
-            width: CARD_WIDTH,
-            height: STEP_CARD_HEIGHT,
-          })),
-        ...topGroups.map(group => {
-          const size = coverSizeOf(group.id);
-          return {
-            id: coverId(group.id),
-            parentIds: group.parentIds.map(parentId =>
-              groupIds.has(parentId) ? coverId(parentId) : parentId,
-            ),
-            order: model.layoutOrder(group.id),
-            width: size.width,
-            height: size.height,
-          };
-        }),
-        ...model.decisions
-          .filter(decision => !model.getNode(decision.parentNodeId)?.groupId)
-          .map((decision, index) => ({
-            id: decision.id,
-            parentIds: [decision.parentNodeId],
-            order: plan.length + index,
-            width: CARD_WIDTH,
-            height: 100,
-          })),
-      ];
-      const layout = computeFlowLayout(layoutItems, STEP_CARD_HEIGHT);
-      const positionOf = (id: string, keepPrev: boolean): { x: number; y: number } => {
-        const prevNode = keepPrev && !relayout ? prevById.get(id) : undefined;
-        return prevNode?.position ?? layout.get(id) ?? { x: 60, y: 40 };
-      };
-
-      const rootNode: Node<EditorNodeData> = {
-        id: VIRTUAL_ROOT_ID,
-        type: 'flowPlanNode',
-        position: positionOf(VIRTUAL_ROOT_ID, true),
-        deletable: false,
-        data: {
-          planNode: null,
-          detached: false,
-          onUpdate: () => {},
-          onDelete: () => {},
-          onAddStep: kind => addStep(null, kind),
-          onConfigure: () => {},
-          configuring: false,
-          readOnly,
-          canAddGroup: true,
-          canAddDecision: false,
-          routesThroughDecision: false,
-        },
-      };
-
-      const orderedActiveGroups = [...topGroups, ...nestedGroups];
-      const coverNodes: Node<EditorNodeData>[] = orderedActiveGroups.map(group => {
-        const isCollapsed = collapsed.has(group.id);
-        const size = coverSizeOf(group.id);
-        const parentNode = group.groupId ? coverId(group.groupId) : undefined;
-        const prevNode = prevById.get(coverId(group.id));
-        const samePlacement = prevNode?.parentNode === parentNode;
-        const position = group.groupId
-          ? !relayout && samePlacement && prevNode
-            ? prevNode.position
-            : (internalByGroup.get(group.groupId)?.memberPositions.get(coverId(group.id)) ?? {
-                x: 20,
-                y: 56,
-              })
-          : positionOf(coverId(group.id), samePlacement);
-        return {
-          id: coverId(group.id),
-          type: 'flowGroupNode',
-          position,
-          style: { width: size.width, height: size.height },
-          draggable: !readOnly,
-          selectable: !readOnly,
-          deletable: false,
-          ...(parentNode && { parentNode }),
-          ...(group.groupId && collapsed.has(group.groupId) && { hidden: true }),
-          data: {
-            name: group.name,
-            memberCount: model.descendantMembersOf(group.id).length,
-            color: flowGroupColor(group.id),
-            collapsed: isCollapsed,
-            ...(isCollapsed && {
-              members: model.directEntityIdsInLevelOrder(group.id).map(id => {
-                const member = model.getNode(id);
-                if (member) return { id: member.id, title: member.title };
-                const child = model.getGroup(id)!;
-                return {
-                  id: child.id,
-                  title: child.name,
-                  memberCount: model.descendantMembersOf(child.id).length,
-                };
-              }),
-            }),
-            connectable: !readOnly,
-            ...(validationWarningById.has(group.id) && {
-              validationWarning: validationWarningById.get(group.id)!,
-            }),
-            onToggleCollapse: () => toggleGroupCollapse(group.id),
-            ...(!readOnly &&
-              onGroupsChange && {
-                onRename: (name: string) => renameGroup(group.id, name),
-                onUngroup: () => ungroup(group.id),
-              }),
-          } satisfies FlowGroupNodeData,
-        };
-      });
-
-      const stepNodes: Node<EditorNodeData>[] = plan.map(step => {
-        const inGroup = step.groupId && groupIds.has(step.groupId);
-        const parentNode = inGroup ? coverId(step.groupId!) : undefined;
-        const containingGroup = step.groupId ? model.getGroup(step.groupId) : undefined;
-        const prevNode = prevById.get(step.id);
-        // Positions only carry over within the same coordinate frame
-        const samePlacement = prevNode?.parentNode === parentNode;
-        const position =
-          !relayout && samePlacement && prevNode
-            ? prevNode.position
-            : inGroup
-              ? (internalByGroup.get(step.groupId!)!.memberPositions.get(step.id) ?? {
-                  x: 20,
-                  y: 56,
-                })
-              : positionOf(step.id, samePlacement);
-        return {
-          id: step.id,
-          type: 'flowPlanNode',
-          position,
-          deletable: false,
-          ...(parentNode && { parentNode }),
-          ...(inGroup &&
-            (collapsed.has(step.groupId!) ||
-              (!!containingGroup?.groupId && collapsed.has(containingGroup.groupId))) && {
-              hidden: true,
-            }),
-          data: {
-            planNode: step,
-            detached: detachedIds.has(step.id),
-            onUpdate: (patch: Partial<FlowPlanNode>) => updateStep(step.id, patch),
-            onDelete: () => deleteStep(step.id),
-            onAddStep: (kind: AddStepKind) => addStep(step.id, kind),
-            onConfigure: (): void => {
-              setConfigDecisionId(null);
-              setConfigNodeId(step.id);
-            },
-            configuring: configNodeId === step.id,
-            readOnly,
-            canAddGroup: !step.groupId || !containingGroup?.groupId,
-            canAddDecision:
-              gateOf(step).type === 'form' &&
-              !decisionsRef.current.some(decision => decision.parentNodeId === step.id) &&
-              !plan.some(candidate => candidate.parentIds.includes(step.id)) &&
-              !groupsRef.current.some(group => group.parentIds.includes(step.id)),
-            routesThroughDecision: decisionsRef.current.some(
-              decision => decision.parentNodeId === step.id,
-            ),
-            ...(validationWarningById.has(step.id) && {
-              validationWarning: validationWarningById.get(step.id)!,
-            }),
-          },
-        };
-      });
-
-      const decisionNodes: Node<EditorNodeData>[] = model.decisions.map(decision => {
-        const source = model.getNode(decision.parentNodeId);
-        const sourceGroupId = source?.groupId ?? null;
-        const inGroup = !!sourceGroupId && groupIds.has(sourceGroupId);
-        const parentNode = sourceGroupId && inGroup ? coverId(sourceGroupId) : undefined;
-        const containingGroup = sourceGroupId ? model.getGroup(sourceGroupId) : undefined;
-        const prevNode = prevById.get(decision.id);
-        const samePlacement = prevNode?.parentNode === parentNode;
-        const position =
-          !relayout && samePlacement && prevNode
-            ? prevNode.position
-            : inGroup
-              ? (internalByGroup.get(sourceGroupId ?? '')?.memberPositions.get(decision.id) ?? {
-                  x: 20,
-                  y: 56,
-                })
-              : positionOf(decision.id, samePlacement);
-        return {
-          id: decision.id,
-          type: 'flowDecisionNode',
-          position,
-          deletable: false,
-          ...(parentNode && { parentNode }),
-          ...(sourceGroupId &&
-            inGroup &&
-            (collapsed.has(sourceGroupId) ||
-              (!!containingGroup?.groupId && collapsed.has(containingGroup.groupId))) && {
-              hidden: true,
-            }),
-          data: {
-            decision,
-            onConfigure: (): void => {
-              setConfigNodeId(null);
-              setConfigDecisionId(decision.id);
-            },
-            onDelete: (): void => {
-              onDecisionsChange?.(
-                decisionsRef.current.filter(candidate => candidate.id !== decision.id),
-              );
-              onChange(
-                planNodesRef.current.map(node =>
-                  node.parentIds.includes(decision.id)
-                    ? {
-                        ...node,
-                        parentIds: node.parentIds.filter(parentId => parentId !== decision.id),
-                      }
-                    : node,
-                ),
-              );
-              onGroupsChange?.(
-                groupsRef.current.map(group =>
-                  group.parentIds.includes(decision.id)
-                    ? {
-                        ...group,
-                        parentIds: group.parentIds.filter(parentId => parentId !== decision.id),
-                      }
-                    : group,
-                ),
-              );
-              setConfigDecisionId(null);
-            },
-            configuring: configDecisionId === decision.id,
-            readOnly,
-            ...(validationWarningById.has(decision.id) && {
-              validationWarning: validationWarningById.get(decision.id)!,
-            }),
-          },
-        };
-      });
-
-      return [rootNode, ...coverNodes, ...stepNodes, ...decisionNodes];
-    },
-    [
-      addStep,
-      updateStep,
-      deleteStep,
-      readOnly,
-      detachedIds,
-      configNodeId,
-      onGroupsChange,
-      onChange,
-      renameGroup,
-      ungroup,
-      toggleGroupCollapse,
-      configDecisionId,
-      onDecisionsChange,
-      validationWarningById,
-    ],
-  );
+  const buildGraphNodes = useBuildFlowPlanNodes({
+    planNodesRef,
+    groupsRef,
+    decisionsRef,
+    collapsedGroupsRef,
+    addStep,
+    updateStep,
+    deleteStep,
+    readOnly,
+    detachedIds,
+    configNodeId,
+    setConfigNodeId,
+    ...(onGroupsChange && { onGroupsChange }),
+    onChange,
+    renameGroup,
+    ungroup,
+    toggleGroupCollapse,
+    configDecisionId,
+    setConfigDecisionId,
+    ...(onDecisionsChange && { onDecisionsChange }),
+    validationWarningById,
+  });
 
   useEffect(() => {
     setNodes(prev => buildGraphNodes(prev, prev.length === 0));
@@ -2125,185 +1275,15 @@ export const FlowPlanEditor: React.FC<FlowPlanEditorProps> = ({
     [readOnly, commit, currentState],
   );
 
-  const edges = useMemo((): Edge[] => {
-    const edgeModel = new FlowPlanModel({
-      version: 2,
-      nodes: planNodes,
-      groups,
-      decisions,
-      updatedAt: 0,
-    });
-    const activeGroupIds = edgeModel.activeGroupIds;
-    const displayId = (planId: string): string =>
-      activeGroupIds.has(planId) ? coverId(planId) : planId;
-    const seen = new Set<string>();
-    const list: Edge[] = [];
-    const pushEdge = (
-      source: string,
-      target: string,
-      options?: { sourceHandle?: string; targetHandle?: string; derived?: boolean },
-      label?: string,
-    ): void => {
-      const id = `e-${source}-${options?.sourceHandle ?? 'default'}-${target}-${options?.targetHandle ?? 'default'}`;
-      if (seen.has(id) || source === target) return;
-      seen.add(id);
-      // Entry/exit edges are derived (parentless / no dependents) — nothing to
-      // delete, but their endpoints can be dragged to rewire the member. Real
-      // parent links get a × delete button (see FlowPlanEdge).
-      const derived = options?.derived ?? false;
-      const removable = !readOnly && source !== VIRTUAL_ROOT_ID && !derived;
-      list.push({
-        id,
-        source,
-        target,
-        ...(options?.sourceHandle ? { sourceHandle: options.sourceHandle } : {}),
-        ...(options?.targetHandle ? { targetHandle: options.targetHandle } : {}),
-        // Derived entry/exit edges keep the default renderer; real links use the
-        // custom edge so they can be deleted with a click.
-        type: derived ? 'default' : 'flowPlan',
-        deletable: removable,
-        updatable: !readOnly && !derived,
-        interactionWidth: 10,
-        ...(removable && {
-          data: { onDelete: (): void => deleteEdgeLink(source, target) },
-        }),
-        markerEnd: {
-          type: MarkerType.ArrowClosed,
-          color: '#6276be',
-          width: 18,
-          height: 18,
-        },
-        style: { stroke: '#94a3b8', strokeWidth: 1.5 },
-        ...(label && {
-          label,
-          labelStyle: {
-            fill: 'hsl(var(--flow-decision-fg))',
-            fontSize: 10,
-            fontWeight: 600,
-          },
-          labelShowBg: true,
-          labelBgStyle: {
-            fill: 'hsl(var(--flow-decision-bg))',
-            stroke: 'hsl(var(--flow-decision-border))',
-            strokeWidth: 1,
-          },
-          labelBgPadding: [6, 4],
-          labelBgBorderRadius: 6,
-        }),
-      });
-    };
-    const externalHandles = (
-      source: string,
-      target: string,
-    ): { sourceHandle?: string; targetHandle?: string } => ({
-      ...(isCoverId(source) && { sourceHandle: GROUP_OUTPUT_HANDLE }),
-      ...(isCoverId(target) && { targetHandle: GROUP_INPUT_HANDLE }),
-    });
-
-    for (const step of planNodes) {
-      if (step.groupId && activeGroupIds.has(step.groupId)) {
-        const containingGroup = edgeModel.getGroup(step.groupId);
-        if (
-          collapsedGroups.has(step.groupId) ||
-          (!!containingGroup?.groupId && collapsedGroups.has(containingGroup.groupId))
-        )
-          continue;
-        for (const parentId of step.parentIds) {
-          if (!decisions.some(decision => decision.id === parentId))
-            pushEdge(displayId(parentId), step.id, externalHandles(displayId(parentId), step.id));
-        }
-        continue;
-      }
-      // Detached steps draw no implicit root edge — they float until rewired
-      const sources =
-        step.parentIds.length > 0
-          ? step.parentIds.map(displayId)
-          : detachedIds.has(step.id)
-            ? []
-            : [VIRTUAL_ROOT_ID];
-      for (const source of sources) {
-        if (!decisions.some(decision => decision.id === source))
-          pushEdge(source, step.id, externalHandles(source, step.id));
-      }
-    }
-    for (const group of groups) {
-      if (!activeGroupIds.has(group.id)) continue;
-      if (group.groupId && collapsedGroups.has(group.groupId)) continue;
-      const sources =
-        group.parentIds.length > 0
-          ? group.parentIds.map(displayId)
-          : detachedIds.has(group.id)
-            ? []
-            : group.groupId
-              ? [coverId(group.groupId)]
-              : [VIRTUAL_ROOT_ID];
-      for (const source of sources) {
-        if (!decisions.some(decision => decision.id === source))
-          pushEdge(
-            source,
-            coverId(group.id),
-            group.groupId && group.parentIds.length === 0
-              ? {
-                  sourceHandle: GROUP_ENTRY_HANDLE,
-                  targetHandle: GROUP_INPUT_HANDLE,
-                  derived: true,
-                }
-              : externalHandles(source, coverId(group.id)),
-          );
-      }
-      if (collapsedGroups.has(group.id)) continue;
-      // Internal flow: cover top fans into entry members, terminal members
-      // fan into the cover bottom — the full path through the group is
-      // visible, and dragging an endpoint rewires the member.
-      const members = planNodes.filter(step => step.groupId === group.id);
-      const terminalIds = new Set(edgeModel.terminalEntityIdsOf(group.id));
-      for (const member of members) {
-        if (member.parentIds.length === 0) {
-          pushEdge(coverId(group.id), member.id, {
-            sourceHandle: GROUP_ENTRY_HANDLE,
-            derived: true,
-          });
-        }
-        if (terminalIds.has(member.id)) {
-          pushEdge(member.id, coverId(group.id), {
-            targetHandle: GROUP_EXIT_HANDLE,
-            derived: true,
-          });
-        }
-      }
-      for (const child of edgeModel.childGroupsOf(group.id)) {
-        if (terminalIds.has(child.id)) {
-          pushEdge(coverId(child.id), coverId(group.id), {
-            sourceHandle: GROUP_OUTPUT_HANDLE,
-            targetHandle: GROUP_EXIT_HANDLE,
-            derived: true,
-          });
-        }
-      }
-    }
-    for (const decision of decisions) {
-      pushEdge(decision.parentNodeId, decision.id, {
-        targetHandle: 'decision',
-        derived: true,
-      });
-      for (const route of decision.routes) {
-        if (!route.targetId) continue;
-        const target = displayId(route.targetId);
-        pushEdge(
-          decision.id,
-          target,
-          {
-            sourceHandle: route.key,
-            ...(isCoverId(target) && { targetHandle: GROUP_INPUT_HANDLE }),
-            derived: true,
-          },
-          route.label,
-        );
-      }
-    }
-    return list;
-  }, [planNodes, groups, decisions, collapsedGroups, detachedIds, readOnly, deleteEdgeLink]);
-
+  const edges = useFlowPlanEdges({
+    planNodes,
+    groups,
+    decisions,
+    collapsedGroups,
+    detachedIds,
+    readOnly,
+    deleteEdgeLink,
+  });
   const configNode = useMemo(
     () => (configNodeId ? (planNodes.find(n => n.id === configNodeId) ?? null) : null),
     [planNodes, configNodeId],
@@ -2399,7 +1379,7 @@ export const FlowPlanEditor: React.FC<FlowPlanEditorProps> = ({
             onEdgeUpdateEnd={onEdgeUpdateEnd}
             onNodeDragStop={onNodeDragStop}
             onSelectionEnd={finishGroupSelection}
-            nodeTypes={NODE_TYPES}
+            nodeTypes={FLOW_PLAN_NODE_TYPES}
             edgeTypes={EDGE_TYPES}
             nodesConnectable={!readOnly}
             edgesUpdatable={!readOnly}
@@ -2568,218 +1548,31 @@ export const FlowPlanEditor: React.FC<FlowPlanEditorProps> = ({
         />
       )}
       {configNode && !readOnly && (
-        <aside className='w-[360px] flex-shrink-0 h-full border-l border-border bg-muted/40 flex flex-col overflow-y-auto'>
-          <div className='flex items-center justify-between px-4 py-3 border-b border-border bg-muted/60'>
-            <span className='text-[11px] font-semibold text-foreground uppercase tracking-wide truncate'>
-              {configNode.title || 'Step'}
-            </span>
-            <button
-              type='button'
-              onClick={() => setConfigNodeId(null)}
-              data-track-category='flow_plan_editor'
-              data-track-name='close_gate_config'
-              className='p-1.5 rounded-md hover:bg-muted text-muted-foreground transition-colors'
-            >
-              <X size={13} />
-            </button>
-          </div>
-          <div className='flex flex-col gap-4 px-4 py-4'>
-            <div>
-              <p className='text-[10px] font-semibold text-muted-foreground uppercase tracking-[0.5px] mb-2'>
-                Default assignee
-              </p>
-              <StepAssigneePicker
-                value={draftAssignedTo}
-                onChange={userId => setDraftAssignedTo(userId)}
-              />
-              <p className='text-[10px] text-muted-foreground mt-1.5 leading-[14px]'>
-                Every run assigns this step&apos;s ticket to this user.
-              </p>
-            </div>
-
-            <div className='border-t border-border pt-4'>
-              <p className='text-[10px] font-semibold text-muted-foreground uppercase tracking-[0.5px] mb-2'>
-                This step waits for
-              </p>
-              <div className='grid grid-cols-2 gap-1.5'>
-                <button
-                  type='button'
-                  onClick={() => setDraftGateType('confirmation')}
-                  data-track-category='flow_plan_editor'
-                  data-track-name='gate_type_confirmation'
-                  className={`flex items-center justify-center gap-1.5 py-1.5 rounded-lg border text-[11px] font-medium transition-colors ${
-                    draftGate.type === 'confirmation'
-                      ? 'bg-[#6276be] border-[#6276be] text-white'
-                      : 'bg-background border-border text-muted-foreground hover:bg-muted'
-                  }`}
-                >
-                  <UserCheck size={12} />
-                  Confirmation
-                </button>
-                <button
-                  type='button'
-                  onClick={() => setDraftGateType('form')}
-                  data-track-category='flow_plan_editor'
-                  data-track-name='gate_type_form'
-                  className={`flex items-center justify-center gap-1.5 py-1.5 rounded-lg border text-[11px] font-medium transition-colors ${
-                    draftGate.type === 'form'
-                      ? 'bg-[#6276be] border-[#6276be] text-white'
-                      : 'bg-background border-border text-muted-foreground hover:bg-muted'
-                  }`}
-                >
-                  <FileText size={12} />
-                  Form
-                </button>
-              </div>
-            </div>
-
-            <div className={draftGate.type === 'confirmation' ? '' : 'hidden'}>
-              <p className='text-[10px] font-semibold text-muted-foreground uppercase tracking-[0.5px] mb-2'>
-                What needs to be confirmed?
-              </p>
-              <Textarea
-                value={draftConfirmationGate.prompt ?? ''}
-                onChange={e =>
-                  setDraftConfirmationGate({ type: 'confirmation', prompt: e.target.value })
-                }
-                placeholder='e.g. KYC documents verified against the registry'
-                rows={3}
-                data-track-category='flow_plan_editor'
-                data-track-name='input_confirmation_prompt'
-                className='min-h-[72px] text-[12px]'
-              />
-            </div>
-            <div className={draftGate.type === 'form' ? '' : 'hidden'}>
-              <p className='text-[10px] font-semibold text-muted-foreground uppercase tracking-[0.5px] mb-2'>
-                Form to fill
-              </p>
-              {formBuilderMode === 'create' ? (
-                <CreateFormSlideOut
-                  embedded
-                  isOpen={true}
-                  onClose={closeFormBuilder}
-                  onSave={handleCreateForm}
-                  submitLabel='Create & attach'
-                  title='Create Step Form'
-                  {...(projectId && { projectId })}
-                />
-              ) : formBuilderMode === 'edit' && draftFormGate.formId && editingFormData ? (
-                <CreateFormSlideOut
-                  embedded
-                  isOpen={true}
-                  onClose={closeFormBuilder}
-                  onSave={() => {}}
-                  onUpdate={handleUpdateAttachedForm}
-                  formId={draftFormGate.formId}
-                  initialData={editingFormData}
-                  submitLabel='Update form'
-                  title='Edit Step Form'
-                  {...(projectId && { projectId })}
-                />
-              ) : draftFormGate.formId ? (
-                <div className='flex flex-col gap-2.5 rounded-lg border border-border bg-background p-3'>
-                  <div className='flex items-start justify-between gap-2'>
-                    <div className='min-w-0'>
-                      <p className='truncate text-[12px] font-semibold text-foreground'>
-                        {formNameById.get(draftFormGate.formId) ?? 'Form'}
-                      </p>
-                      <p className='text-[10px] text-muted-foreground'>
-                        {attachedFieldRows?.length ?? 0} field
-                        {(attachedFieldRows?.length ?? 0) === 1 ? '' : 's'}
-                      </p>
-                    </div>
-                    <div className='flex items-center gap-1'>
-                      <button
-                        type='button'
-                        title='Edit form'
-                        onClick={() => void handleOpenAttachedForm()}
-                        data-track-category='flow_plan_editor'
-                        data-track-name='edit_gate_form'
-                        className='p-1 rounded hover:bg-muted text-muted-foreground hover:text-foreground transition-colors'
-                      >
-                        <Settings2 size={12} />
-                      </button>
-                      <button
-                        type='button'
-                        title='Detach form'
-                        onClick={() => setDraftFormGate({ type: 'form', formId: '' })}
-                        data-track-category='flow_plan_editor'
-                        data-track-name='remove_gate_form'
-                        className='p-1 rounded hover:bg-red-50 text-muted-foreground hover:text-red-500 transition-colors'
-                      >
-                        <X size={11} />
-                      </button>
-                    </div>
-                  </div>
-                  {attachedFieldRows === undefined ? (
-                    <p className='text-[11px] text-muted-foreground'>Loading fields…</p>
-                  ) : (
-                    <div className='flex flex-col divide-y divide-border'>
-                      {attachedFieldRows.map(row => {
-                        const fieldName = row.globalField?.fieldName ?? row.fieldName;
-                        const fieldType = row.globalField?.fieldType ?? row.fieldType;
-                        const fieldEnum = row.globalField?.fieldEnum ?? row.fieldEnum;
-                        if (!fieldName || !fieldType) return null;
-                        return (
-                          <div key={row.id} className='py-2 first:pt-0 last:pb-0'>
-                            <div className='flex items-start justify-between gap-2'>
-                              <span className='text-[11px] font-medium text-foreground'>
-                                {fieldName}
-                                {!row.isOptional && <span className='text-red-500'> *</span>}
-                              </span>
-                              <span className='shrink-0 text-[10px] text-muted-foreground'>
-                                {getFieldTypeLabel(fieldType)}
-                              </span>
-                            </div>
-                            {Array.isArray(fieldEnum) && fieldEnum.length > 0 && (
-                              <p className='mt-0.5 line-clamp-2 text-[10px] text-muted-foreground'>
-                                {(fieldEnum as string[]).join(' · ')}
-                              </p>
-                            )}
-                          </div>
-                        );
-                      })}
-                    </div>
-                  )}
-                </div>
-              ) : (
-                <TransitionFormPicker
-                  allForms={formOptions}
-                  onCreateForm={() => setFormBuilderMode('create')}
-                  onSelectForm={formId => setDraftFormGate({ type: 'form', formId })}
-                  variant='dashed-button'
-                  triggerLabel='Attach form'
-                />
-              )}
-              <p className='text-[11px] text-muted-foreground mt-2 leading-[16px]'>
-                At runtime the assignee fills this form on the step; submitting it completes the
-                step.
-              </p>
-            </div>
-          </div>
-          <div className='sticky bottom-0 mt-auto flex gap-2 border-t border-border bg-background px-4 py-3'>
-            <Button
-              variant='secondary'
-              size='sm'
-              className='flex-1'
-              onClick={() => setConfigNodeId(null)}
-              data-track-category='flow_plan_editor'
-              data-track-name='cancel_gate_config'
-            >
-              Cancel
-            </Button>
-            <Button
-              size='sm'
-              className='flex-1 bg-[#6276BE] hover:bg-[#5060A0] text-white'
-              disabled={!canSavePanel}
-              onClick={handleSavePanel}
-              data-track-category='flow_plan_editor'
-              data-track-name='save_gate_config'
-            >
-              Save
-            </Button>
-          </div>
-        </aside>
+        <FlowStepConfigPanel
+          configNode={configNode}
+          onClose={() => setConfigNodeId(null)}
+          draftAssignedTo={draftAssignedTo}
+          setDraftAssignedTo={setDraftAssignedTo}
+          draftGate={draftGate}
+          setDraftGateType={setDraftGateType}
+          draftConfirmationGate={draftConfirmationGate}
+          setDraftConfirmationGate={setDraftConfirmationGate}
+          draftFormGate={draftFormGate}
+          setDraftFormGate={setDraftFormGate}
+          formBuilderMode={formBuilderMode}
+          setFormBuilderMode={setFormBuilderMode}
+          closeFormBuilder={closeFormBuilder}
+          handleCreateForm={handleCreateForm}
+          handleOpenAttachedForm={handleOpenAttachedForm}
+          {...(projectId && { projectId })}
+          editingFormData={editingFormData}
+          handleUpdateAttachedForm={handleUpdateAttachedForm}
+          formNameById={formNameById}
+          attachedFieldRows={attachedFieldRows}
+          formOptions={formOptions}
+          canSavePanel={canSavePanel}
+          handleSavePanel={handleSavePanel}
+        />
       )}
     </div>
   );
